Move activity-level buttons away from responders status (#4)
To reduce confusion, move the buttons that control an activity (edit, complete/reactivate, delete) away from the responder status button. Specifically to the right of the title.

Validated visual behavior with long titles.

Note that the height needed to be set for the button stack, otherwise the buttons would expand in height as the title wrapped onto multiple lines. The fixed height matches the font size of the title so that the buttons are as tall as one line of title text.

---------

Co-authored-by: Matt Cosand <mcosand@users.noreply.github.com>