Support per-org favicon (#33)
Adding support for an organization to define a favicon and home screen icon (for when the site is pinned to a home screen on mobile). There are no filesystem changes (e.g. adding images) as the intent is organizations will set the icon to match their publicly facing website (or an image file hosted there).

While the optimal home screen icon solution would require responding with a custom site manifest for Android, by my research it will happily fall back to apple-touch-icon if the site manifest isn't present. I'll come back and add support for the site manifest if this doesn't work well.