Consistently display 24 hour time (#34)
We weren't displaying 24 hour time for the start and end times for activities. This is because they were using date-fns formatRelative method, which displays 12 hour time by default without an obvious way to make it 24 hour time.

Figured out how to make formatRelative display 24 hour time. It does mean that formatRelative is now specific to enUS; however, I feel that's appropriate for the audience of this app for the foreseeable future.

Additionally, removed colons from time display.