Force line endings to LF, even on Windows (#51)
Git on Windows defaults check out files with CRLF line endings and push them with LF line endings for consistency with other platforms. This generally works, but our linter is exclusively looking for LF endings. This means it's impossible to run the linter locally on Windows with the default Git behavior.

This change updates the README and .gitattributes to direct the user to set autocrlf to false for this repository.