Custom Sign In Window (#60)
Adds support for a per-activity early sign-in window. Options are:

    4 hours (current fixed window, and remains the default).
    12 hours
    24 hours
    Unlimited (no restriction; stored as Infinity)

Existing activities in the database will fall back to the default of 4 hours, though it shouldn't really matter.

Field is disabled in the editor if the start date of the mission is not in the future.

I opted to keep the list of options in the same place as the default, as I wanted the default to be defined from the options. Otherwise it's possible for the default value to be changed to an option that doesn't exist. An enum would have helped here, but I didn't want to store the enum in the database as a number leaves more long-term flexibility.